chore: move delete all stuff permission check to policy

+++ root/routes/api.php
@@ -170,9 +170,8 @@ Route::middleware('auth:sanctum')->group(function() {
     Route::get('/user', [UserController::class, 'currentUser']);
     Route::middleware('throttle:1,1')->get('/user-data', [UserController::class, 'userData']);
     Route::patch('users/{user}/roles', [UserController::class, 'setUserRoles']);
-    Route::delete('users/{user}/mods', [UserController::class, 'deleteMods']);
-    Route::delete('users/{user}/discussions', [UserController::class, 'deleteDiscussions']);
-
+    Route::middleware('can:manageMods,user')->delete('users/{user}/mods', [UserController::class, 'deleteMods']);
+    Route::middleware('can:manageDiscussions,user')->delete('users/{user}/discussions', [UserController::class, 'deleteDiscussions']);
     Route::resource('blocked-users', BlockedUserController::class)->except('show', 'update');
     Route::resource('blocked-tags', BlockedTagController::class)->except('show', 'update');
     Route::resource('followed-mods', FollowedModController::class)->except('show');
